Add config, refactor UserServiceImpl

+++ src/main/java/ru/kata/springbootsecuritybootstrap/services/UserService.java
@@ -1,7 +1,8 @@
 package ru.kata.springbootsecuritybootstrap.services;
 
-import org.springframework.security.core.userdetails.User;
+
 import ru.kata.springbootsecuritybootstrap.models.Role;
+import ru.kata.springbootsecuritybootstrap.models.User;
 
 import java.util.List;
 import java.util.Set;
@@ -11,15 +12,15 @@ public interface UserService {
 
     void save(User user);
 
-    void delete(Long id);
+    void delete(Integer id);
 
-    User getOne(Long id);
+    User getOne(Integer id);
 
-    void update(Long id, User user);
+    void update(Integer id, User user);
 
     User oneUser();
 
     User createUser(User user, Set<Role> roles);
 
-    User updateUser(User user, Set<Role> roles, Long id);
+    User updateUser(User user, Set<Role> roles, Integer id);
 }
